Stop overriding extractCreativeAndSignature in impls
