+++ extensions/amp-a4a/0.1/test/utils.js
@@ -15,13 +15,6 @@
  */
 
 import {AmpA4A} from '../amp-a4a';
-import {base64UrlDecodeToBytes} from '../../../../src/utils/base64';
-
-/** @type {string} @private */
-export const SIGNATURE_HEADER = 'X-TestSignatureHeader';
-
-/** @type {string} @private */
-export const SIZE_HEADER = 'X-CreativeSize';
 
 /** @type {string} @private */
 export const TEST_URL = 'http://iframe.localhost:' + location.port +
@@ -34,18 +27,6 @@ export class MockA4AImpl extends AmpA4A {
 
   updatePriority() {
     // Do nothing.
-  }
-
-  extractCreativeAndSignature(responseArrayBuffer, responseHeaders) {
-    const sizeArr = responseHeaders.has(SIZE_HEADER) ?
-        responseHeaders.get(SIZE_HEADER).split('x') : null;
-    const size = sizeArr ? {width: sizeArr[0], height: sizeArr[1]} : null;
-    return Promise.resolve({
-      creative: responseArrayBuffer,
-      signature: responseHeaders.has(SIGNATURE_HEADER) ?
-          base64UrlDecodeToBytes(responseHeaders.get(SIGNATURE_HEADER)) : null,
-      size,
-    });
   }
 
   getFallback() {
