+++ extensions/amp-ad-network-cloudflare-impl/0.1/amp-ad-network-cloudflare-impl.js
@@ -15,19 +15,9 @@
  */
 
 import {AmpA4A} from '../../amp-a4a/0.1/amp-a4a';
-import {base64UrlDecodeToBytes} from '../../../src/utils/base64';
-import {dev} from '../../../src/log';
 import {startsWith} from '../../../src/string';
 import {assertAbsoluteHttpOrHttpsUrl} from '../../../src/url';
 import {NETWORKS} from './vendors';
-
-/**
- * Header that will contain Cloudflare generated signature
- *
- * @type {string}
- * @private
- */
-const AMP_SIGNATURE_HEADER = 'X-AmpAdSignature';
 
 /**
  * This is a minimalistic AmpA4A implementation that primarily gets an Ad
@@ -122,29 +112,6 @@ export class AmpAdNetworkCloudflareImpl extends AmpA4A {
     }
 
     return url;
-  }
-
-  /**
-   * Extract creative and signature from a Cloudflare signed response.
-   *
-   * Note: Invalid A4A content will NOT have a signature, which will automatically
-   *   cause the A4A processing to render it within a cross domain frame.
-   *
-   * @override
-   */
-  extractCreativeAndSignature(responseText, responseHeaders) {
-    let signature = null;
-    try {
-      if (responseHeaders.has(AMP_SIGNATURE_HEADER)) {
-        signature =
-          base64UrlDecodeToBytes(dev().assertString(
-              responseHeaders.get(AMP_SIGNATURE_HEADER)));
-      }
-    } finally {
-      return Promise.resolve(/** @type {!../../../extensions/amp-a4a/0.1/amp-a4a.AdResponseDef} */
-        ({creative: responseText, signature})
-      );
-    }
   }
 }
